bug fixed

+++ include/timer.h
@@ -68,7 +68,7 @@ extern void TMR1_init(unsigned int reload , void (*func)());
 void TMR1_ISR(void);
 void TMR0_ISR(void);
 
-UINT32 GetAppTime(void);
+UINT16 GetAppTime(void);
 void ResetAppTime(void);
